Implemented Implemented, Toaster, Button disable, and view setting during and after api call. Error handling of stripe.

+++ TodoApp/Domain/Commands/Stripe/PaymentDeductCommand.cs
@@ -10,6 +10,6 @@ namespace Fistix.Training.Domain.Commands.Stripe
     //public string Email { get; set; }
     public string StripeCustomerId { get; set; }
     public string CustomerId { get; set; }
-    public long Amount { get; set; }
+    public int Amount { get; set; }
   }
 }
