+++ TodoApp/Domain/Commands/Stripe/SampleCheckoutCommand.cs
@@ -8,7 +8,7 @@ namespace Fistix.Training.Domain.Commands.Stripe
   public class SampleCheckoutCommand : IRequest<SampleCheckoutCommandResult>
   {
     public string Email { get; set; }
-    public long Amount { get; set; }
+    public int Amount { get; set; }
     public string ProductName { get; set; }
   }
 }
